% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{daily_summary}
\alias{daily_summary}
\title{Daily Summary Feed API}
\usage{
daily_summary(key = get_key())
}
\arguments{
\item{key}{Your API key (retrieved automatically if stored in \emph{.Renviron})  OR \code{access_token} - the access token from the Oauth2 Connection}
}
\value{
\code{(list)}
}
\description{
GET data from the \href{https://www.rescuetime.com/anapi/setup/documentation#daily-summary-feed-reference}{Daily Summary Feed API}
}
