% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{end_focustime}
\alias{end_focustime}
\title{FocusTime Trigger API}
\usage{
end_focustime(key = get_key())
}
\arguments{
\item{key}{Your API key (retrieved automatically if stored in \emph{.Renviron})  OR \code{access_token} - the access token from the Oauth2 Connection}
}
\value{
No warning if successful
}
\description{
POST data to the \href{https://www.rescuetime.com/anapi/setup/documentation#focustime-trigger-reference}{FocusTime Trigger API}
}
