% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{highlights}
\alias{highlights}
\title{Daily Highlights Feed API}
\usage{
highlights(key = get_key())
}
\arguments{
\item{key}{Your API key (retrieved automatically if stored in \emph{.Renviron})  OR \code{access_token} - the access token from the Oauth2 Connection}
}
\value{
\code{(list)}
}
\description{
GET data from the \href{https://www.rescuetime.com/anapi/setup/documentation#highlights-feed-reference}{Daily Highlights Feed API}
}
