% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{highlights_post}
\alias{highlights_post}
\title{Daily Highlights Post API}
\usage{
highlights_post(key = get_key())
}
\arguments{
\item{key}{Your API key (retrieved automatically if stored in \emph{.Renviron})  OR \code{access_token} - the access token from the Oauth2 Connection}
}
\value{
No warning if successful
}
\description{
POST data to the \href{https://www.rescuetime.com/anapi/setup/documentation#highlights-post-reference}{Daily Highlights Post API}
}
