% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{offline}
\alias{offline}
\title{Offline Time POST API}
\usage{
offline(
  key = get_key(),
  start_time = Sys.time() - lubridate::dminutes(5),
  end_time = 5,
  activity_name = c("Break"),
  activity_details = NULL
)
}
\arguments{
\item{start_time}{\itemize{
\item A string representing the date/time the for the start of the offline time block. This should be in the format of ‘YYYY-MM-DD HH:MM:SS’, but a unix timestamp is also acceptable.
}}

\item{end_time}{\itemize{
\item (or duration) Either a string representing the date/time the for the end of the offline time block, OR an integer representing the duration of the offline time block in minutes.
}}

\item{activity_name}{\itemize{
\item A 255 character or shorter string containing the text that will be entered as the name of the activity (e.g. "Meeting", "Driving", "Sleeping", etc).
}}

\item{activity_details}{\itemize{
\item \emph{Optional} A 255 character or shorter string containing the text that will be entered as the details of the named activity.
}}
}
\value{
No warning if successful
}
\description{
GET data from the \href{https://www.rescuetime.com/anapi/setup/documentation#offline-time-reference}{Offline Time POST API}
}
