% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup Rescuetime by adding the RESCUETIME_KEY environmental variable}
\usage{
setup(key = "[Your Key]", .Renviron_path = ".Renviron")
}
\arguments{
\item{key}{Your key. See \href{https://www.rescuetime.com/anapi/manage}{API Key Management}}

\item{.Renviron_path}{Path to the \emph{.Renviron} file to modify. Defaults to current directory. If one does not exist it will be created.}
}
\value{
Messages or warnings as is suitable
}
\description{
Setup Rescuetime by adding the RESCUETIME_KEY environmental variable
}
\examples{
setup(
"fake_key",
# to set a user level environmental variable (otherwise the default working_directory is used)
.Renviron_path = "~/.Renviron"
)
}
